Added new type Chat (as ChatInfo), enum ChatType, blank for ChatPhoto (as ChatPhotoInfo). Not finished. Not tested.
