+++ NetTelebot.Tests/NetTelebot.Tests.csproj
@@ -69,6 +69,7 @@
     <Compile Include="MessageInfoParserTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TelegramBotClientTest.cs" />
+    <Compile Include="TelegramRealBotClientTest.cs" />
     <Compile Include="TypeTestObject\AudioInfoObject.cs" />
     <Compile Include="TypeTestObject\ContactInfoObject.cs" />
     <Compile Include="TypeTestObject\DocumentInfoObject.cs" />
