+++ NetTelebot/ChatPhotoInfo.cs
@@ -1,0 +1,12 @@
+﻿namespace NetTelebot
+{
+    public class ChatPhotoInfo
+    {
+        internal ChatPhotoInfo()
+        {
+            
+        }
+
+        //todo add type
+    }
+}
