+++ NetTelebot/ChatType.cs
@@ -1,0 +1,28 @@
+﻿namespace NetTelebot
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public enum ChatType
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        @private,
+
+        /// <summary>
+        /// 
+        /// </summary>
+        group,
+
+        /// <summary>
+        /// 
+        /// </summary>
+        supergroup,
+
+        /// <summary>
+        /// 
+        /// </summary>
+        channel
+    }
+}
